refact(validator): move validator into request package
